requirements.txt
anthropic>=0.40.0
pydantic>=2.0
typer>=0.9.0
rich>=13.0
httpx>=0.25.0
pyyaml>=6.0
python-dotenv>=1.0.0
asyncio>=3.4.3
